+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [[v0.1.1]](https://github.com/mlange-42/ark-pixel/compare/v0.1.0...v0.1.1)
 
 - Number of entities per archetype is displayed in `Monitor` (#13)
+- Reserved and used memory are both displayed in `Monitor` (#14)
 
 ## [[v0.1.0]](https://github.com/mlange-42/ark-pixel/commits/v0.1.0/)